Re #39 Refactor commandline arg parsing to separate file
